*Description*:
This change adds a performance benchmark test for the HTTP header map implementation.

*Risk Level*: low

*Testing*:
```
$ bazel build -c opt //test/common/http/...
[build output omitted for clarity]
$ bazel-bin/test/common/http/header_map_impl_speed_test
$ bazel-bin/test/common/http/header_map_impl_speed_test
2019-01-20 17:49:24
Running bazel-bin/test/common/http/header_map_impl_speed_test
Run on (12 X 3500 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 262K (x6)
  L3 Unified 12582K (x1)
--------------------------------------------------------------------
Benchmark                             Time           CPU Iterations
--------------------------------------------------------------------
HeaderMapImplCreate                  15 ns         15 ns   47149160
HeaderMapImplSetReference/0         139 ns        139 ns    5166128
HeaderMapImplSetReference/1         141 ns        141 ns    4321841
HeaderMapImplSetReference/10        208 ns        208 ns    3363412
HeaderMapImplSetReference/50        441 ns        441 ns    1625635
HeaderMapImplGet/0                   15 ns         15 ns   47585382
HeaderMapImplGet/1                   21 ns         21 ns   33841602
HeaderMapImplGet/10                  75 ns         75 ns    9267963
HeaderMapImplGet/50                 320 ns        320 ns    2185833
HeaderMapImplGetInline/0              2 ns          2 ns  366732155
HeaderMapImplGetInline/1              2 ns          2 ns  369582319
HeaderMapImplGetInline/10             2 ns          2 ns  367703064
HeaderMapImplGetInline/50             2 ns          2 ns  339792629
HeaderMapImplSetInline/0              8 ns          8 ns   83102822
HeaderMapImplSetInline/1              8 ns          8 ns   82857887
HeaderMapImplSetInline/10             8 ns          8 ns   82959030
HeaderMapImplSetInline/50             8 ns          8 ns   82152875
```

*Docs Changes*: N/A
*Release Notes*: N/A

Signed-off-by: Brian Pane <brianp+github@brianp.net>