@jmarantz 2019-01-21
Love it! Could we also test performance of the other methods of HeaderMap?  lookup() and remove() seem interesting. Also iterate().

I'm curious: are you thinking of an alternate implementation to benchmark?