@brian-pane 2019-01-21
Sure, I'll add benchmarks for those other methods.

With regard to alternate implementations, I've been thinking for a long time about trying the same header map [design](https://github.com/facebook/proxygen/blob/master/proxygen/lib/http/HTTPHeaders.h#L46) that we used in Proxygen:
- Generate a consistent hash function that maps the commonly used header keys to values in the [0,254] range.
- Represent the headers as an array of keys (1 byte each) and a parallel array of values (strings).
- To add a header, append it to the two arrays.
- To retrieve the value of a common key, scan the key array to find all entries with the key's hash value and return the corresponding entries from the value array.
- Headers that don't match the common keys get stored in a separate map.

The main performance tricks in that design are:
- Cache locality: the array of 1-byte keys often will fit in a single cache line.
- Parallelism: since the common keys are encoded as single bytes, the search for a common key in the _n_ headers can be done with just _n_/16 or _n_/32 comparisons and conditional branch operations, depending on the SIMD register width. That brings the O(_n_) search time closer to O(1) in the common case. In the worst case, it's still O(_n_), so the design relies on the proxy to impose a limit on the header size of incoming messages (which it needs to do anyway, to protect against malicious requests using too much memory).