@jmarantz 2019-01-21
Yeah that was my general thought too: in the past when I've wanted a fast lexer, 'gperf' has been my go-to tool for generating perfect hash functions. Having said that the trie in this impl seems pretty fast relative to other things that might be going on in the proxy. Is there anything something telling you headers are a performance bottleneck?

In ay case this work is great for quantifying the potential benefit. Ping when ready for another look.