@brian-pane 2019-01-21
In CPU profiles, I've seen the HeaderMapImpl methods collectively accounting for 1-2% of total CPU. I don't have a good way to quantify the impact of HeaderMapImpl's cache footprint on the performance of other things in the request path, though. My hypothesis is that constructing a HeaderMapImpl - which involves initializing about half a KB of null pointers, for all the O(1)-time fields - will push enough other things out of L1D to cause subsequent slowdowns elsewhere, but the only way to find out for sure will be to load-test alternate designs.