@mattklein123 2019-01-22
The proposed alternate implementation sounds pretty neat and it may very well be faster. One thing I will point out about the trie is that it's a very stupid implementation that makes many more node comparisons than needed. I think if it were implemented as a compressed trie with character skipping (using intrinsics for comparison) the number of comparisons would go way down. Not sure how much that would help on the benchmarks but throwing that out there.